requirements.txt

streamlit
requests
gspread
google-auth
anthropic
python-dotenv
openai